requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
boto3==1.35.0
botocore==1.35.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
fastdtw==0.3.4
Flask==3.1.1
flask-cors==6.0.0
flatbuffers==25.2.10
fonttools==4.58.5
gotrue==2.12.3
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
openai==1.82.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
postgrest==1.1.1
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
realtime==2.5.3
redis==5.0.0
requests==2.32.3
s3transfer==0.10.0
scipy==1.15.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
storage3==0.12.0
StrEnum==0.4.15
stripe==12.3.0
supabase==2.16.0
supafunc==0.10.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
